Add files via upload
updated playback function and pni Data class to prevent 'garbage collection' of my plots.  however i think i need to turn it 
into a list of notes that i append and pop.  i feel like there is a better way, but since this is not a 'procedural program' rather a 'living loop,'
it may be the only way i now know .  it only plays one note repeatedly currently
bc