Add files via upload
I re-structured structure.py to account for the 'multi-pass approach' since I no longer use a forLoop in it.  much of the original code has been commented out.  
sometimes it works correctly, but sometimes it gets stuck on the wrong gate 'generate notes' and doesn't playback.  why is this?  
bc