Add files via upload
i am about to completely revamp my note creation in 'generate_notes.py'.  I will completely remove the 'gni.createNote() ' function and simply create a 
Note object outright in the 'generate_notes.py' module.  it will be a long module.  but i think that is okay for now.
i also added a 'select_image()' function to 'generate_notes.py' and recently added an indentation to 'generate()' function so as to allow for the creation of different patterns and rules for the different fingers/polyphonies. i also began to update state locally in the files such as 'plot_notes' and added a 'plot_copy' of the song that i will populate and then depopulate to check progress in the debug_log.  
bc