Add files via upload
general update.  focused on performance.  i think my debug_log with large text_handler was messing things up.  when i commented it out for the debug_log page
it seemed to improve performance, but i seem to still be slowing down and crackling sound perhaps due to the large number of channels i imposed in the mainLoop (50 channels).  I plan to reduce channels to 24 and see if this improves the situation.   I also need to think about adding rests at some point which can potentially further reduce the number of active channels.  
bc